refactor: update user entity folder structor
