--- src/entities/errors/index.ts
@@ -1,3 +1,0 @@
-export * from '@/entities/errors/invalid-name-error';
-export * from '@/entities/errors/invalid-email-error';
-export * from '@/entities/errors/invalid-password-error';
