--- src/entities/interfaces/index.ts
@@ -1,1 +1,0 @@
-export * from '@/entities/interfaces/user-create-interface';
