+++ src/entities/user/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from '@/entities/user/interfaces/user-create-interface';
