+++ src/entities/value-object/errors/index.ts
@@ -1,0 +1,3 @@
+export * from '@/entities/value-object/errors/invalid-name-error';
+export * from '@/entities/value-object/errors/invalid-email-error';
+export * from '@/entities/value-object/errors/invalid-password-error';
