+++ src/entities/value-object/errors/invalid-email-error.ts
