+++ src/entities/value-object/errors/invalid-name-error.ts
