+++ src/entities/value-object/errors/invalid-password-error.ts
